requirements.txt
# LSEG-DATA and DataStreamPy Dependencies

# DataStream Web Service Python library
DatastreamDSWS>=1.0.0

# LSEG Data Library for Python
lseg-data>=1.0.0

# Common dependencies
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.5.0
python-dotenv>=0.19.0
